+++ engine/table/src/main/java/io/deephaven/engine/table/impl/select/vectorchunkfilter/CharVectorChunkFilter.java
@@ -6,13 +6,15 @@ package io.deephaven.engine.table.impl.select.vectorchunkfilter;
 import io.deephaven.chunk.*;
 import io.deephaven.chunk.attributes.Values;
 import io.deephaven.engine.primitive.value.iterator.ValueIteratorOfChar;
-import io.deephaven.engine.rowset.chunkattributes.OrderedRowKeys;
-import io.deephaven.util.mutable.MutableInt;
 import io.deephaven.vector.CharVector;
 
 import java.util.function.IntConsumer;
 import java.util.function.IntPredicate;
 
+/**
+ * A wrapper that extracts elements from a CharVector, returning true for the CharVector if any elements are matched by
+ * the wrapped chunk filter.
+ */
 class CharVectorChunkFilter extends VectorChunkFilter {
     final WritableCharChunk<? extends Values> temporaryValues;
 
@@ -43,7 +45,7 @@ class CharVectorChunkFilter extends VectorChunkFilter {
                     srcPos.set(fillPos, indexOfVector);
                     temporaryValues.set(fillPos++, element);
                     if (fillPos == chunkSize) {
-                        final long lastMatch = flushMatches(matchConsumer, fillPos, temporaryValues);
+                        final int lastMatch = flushMatches(matchConsumer, fillPos, temporaryValues);
                         fillPos = 0;
                         if (lastMatch == indexOfVector) {
                             break;
